requirements.txt
typing
jpype
re
json
numpy
gensim.models
os
random
flask
flask_cors